+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Implement `Clone` and `Copy` for validation error types.
 * Let an error type contain source string for conversion from owned string.
 * Add `shrink_to_fit()` methods for `types::iri::*String` types.
+* Add `set_fragment()` methods for `types::iri::*String` types
+  (except for `AbsoluteIriString`).
 * Add `as_str()` method for `types::iri::*Str` types.
 * Add `types::iri::IriFragment{Str,String}` type.
 * Move `fragment()` from `IriStr` to `IriReferenceStr`.
@@ -15,6 +17,8 @@
 
 #### Added
 * Add `shrink_to_fit()` methods for `types::iri::*String` types.
+* Add `set_fragment()` methods for `types::iri::*String` types
+  (except for `AbsoluteIriString`).
 * Add `as_str()` method for `types::iri::*Str` types.
 * Add `types::iri::IriFragment{Str,String}` type.
     + This represents fragment part of an IRI.